agregando sp huesped, bloqueando textos y botones

+++ ProyectoFinal/Huesped.cs
@@ -11,7 +11,7 @@ namespace ProyectoFinal
         private string idHuesped;
         private string nombresHuesped;
         private string apellidosHuesped; 
-        private int numIdentificacionHuesped;
+        private long numIdentificacionHuesped;
         private string direccionHuesped;
         private string emailHuesped;
         private int telefonoHuesped;
@@ -19,7 +19,7 @@ namespace ProyectoFinal
 
         
 
-        public Huesped(string idHuesped, string nombresHuesped, string apellidosHuesped, int numIdentificacionHuesped, string direccionHuesped, string emailHuesped, int telefonoHuesped, string tipoHuesped)
+        public Huesped(string idHuesped, string nombresHuesped, string apellidosHuesped, long numIdentificacionHuesped, string direccionHuesped, string emailHuesped, int telefonoHuesped, string tipoHuesped)
         {
             this.IdHuesped = idHuesped;
             this.NombresHuesped = nombresHuesped;
@@ -34,7 +34,7 @@ namespace ProyectoFinal
         public string IdHuesped { get => idHuesped; set => idHuesped = value; }
         public string NombresHuesped { get => nombresHuesped; set => nombresHuesped = value; }
         public string ApellidosHuesped { get => apellidosHuesped; set => apellidosHuesped = value; }
-        public int NumIdentificacionHuesped { get => numIdentificacionHuesped; set => numIdentificacionHuesped = value; }
+        public long NumIdentificacionHuesped { get => numIdentificacionHuesped; set => numIdentificacionHuesped = value; }
         public string DireccionHuesped { get => direccionHuesped; set => direccionHuesped = value; }
         public string EmailHuesped { get => emailHuesped; set => emailHuesped = value; }
         public int TelefonoHuesped { get => telefonoHuesped; set => telefonoHuesped = value; }
